requirements.txt
# requirements.txt (core libraries minus torch)
# pip install torch==2.6.0+cu126 torchvision==0.15.0+cu126 torchaudio==2.6.0 --extra-index-url https://download.pytorch.org/whl/cu126
# first install torch and torchvision

timm>=0.9.0
transformers>=4.30.0
numpy<2.0
pandas>=2.0.0
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
grad-cam>=1.4.0
tensorboard>=2.12.0
tensorboardX
facenet-pytorch>=2.5.2
pyyaml
json_tricks
yacs
tensorwatch
ptflops
einops
pytest

fastapi
uvicorn
opencv-python-headless
python multipart